Refactor native user sign up and login into model
